+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/NoSuchControlException.java
@@ -1,0 +1,8 @@
+package hu.gds.ldap4j.ldap;
+
+import java.io.Serial;
+
+public class NoSuchControlException extends RuntimeException {
+    @Serial
+    private static final long serialVersionUID=0L;
+}